+++ main.js
@@ -1,1 +1,235 @@
-var _0xd2a4=["\x41\x49\x7A\x61\x53\x79\x41\x59\x32\x2D\x55\x2D\x77\x72\x41\x72\x57\x45\x6E\x78\x55\x69\x54\x41\x65\x6F\x54\x62\x72\x4A\x49\x30\x5F\x71\x6B\x43\x75\x54\x73","\x74\x79\x70\x65\x62\x65\x61\x73\x74\x2D\x66\x39\x63\x38\x30\x2E\x66\x69\x72\x65\x62\x61\x73\x65\x61\x70\x70\x2E\x63\x6F\x6D","\x68\x74\x74\x70\x73\x3A\x2F\x2F\x74\x79\x70\x65\x62\x65\x61\x73\x74\x2D\x66\x39\x63\x38\x30\x2E\x66\x69\x72\x65\x62\x61\x73\x65\x69\x6F\x2E\x63\x6F\x6D","\x74\x79\x70\x65\x62\x65\x61\x73\x74\x2D\x66\x39\x63\x38\x30","\x74\x79\x70\x65\x62\x65\x61\x73\x74\x2D\x66\x39\x63\x38\x30\x2E\x61\x70\x70\x73\x70\x6F\x74\x2E\x63\x6F\x6D","\x39\x36\x38\x37\x31\x36\x33\x33\x36\x30\x39\x39","\x31\x3A\x39\x36\x38\x37\x31\x36\x33\x33\x36\x30\x39\x39\x3A\x77\x65\x62\x3A\x64\x34\x65\x61\x30\x37\x66\x30\x34\x33\x61\x35\x39\x37\x36\x31\x39\x38\x62\x65\x64\x36","\x47\x2D\x32\x42\x4A\x51\x39\x56\x42\x33\x52\x59","\x69\x6E\x69\x74\x69\x61\x6C\x69\x7A\x65\x41\x70\x70","\x66\x69\x72\x65\x73\x74\x6F\x72\x65","","\x66\x61\x64\x65\x49\x6E","\x23\x67\x61\x6D\x65\x69\x6E\x70\x75\x74","\x70\x6C\x61\x63\x65\x68\x6F\x6C\x64\x65\x72","\x67\x61\x6D\x65\x69\x6E\x70\x75\x74","\x67\x65\x74\x45\x6C\x65\x6D\x65\x6E\x74\x42\x79\x49\x64","\x53\x74\x61\x72\x74\x20\x74\x79\x70\x69\x6E\x67\x20\x74\x6F\x20\x70\x6C\x61\x79\x2E\x2E\x2E","\x76\x61\x6C\x75\x65","\x69\x6E\x6E\x65\x72\x48\x54\x4D\x4C","\x67\x61\x6D\x65\x74\x65\x78\x74","\x20","\x73\x70\x6C\x69\x74","\x6C\x65\x6E\x67\x74\x68","\x64\x69\x73\x61\x62\x6C\x65\x64","\x77\x70\x6D","\x57\x50\x4D\x3A\x20\x3F","\x66\x6F\x6E\x74\x53\x69\x7A\x65","\x73\x74\x79\x6C\x65","\x73\x68\x69\x66\x74","\x6A\x6F\x69\x6E","\x66\x61\x64\x65\x4F\x75\x74","\x70\x75\x73\x68","\x73\x75\x62\x73\x74\x72","\x57\x50\x4D\x3A\x20","\x62\x65\x73\x74","\x59\x6F\x75\x72\x20\x62\x65\x73\x74\x3A\x20","\x61\x76\x65\x72\x61\x67\x65","\x59\x6F\x75\x72\x20\x61\x76\x65\x72\x61\x67\x65\x3A\x20","\x34\x72\x65\x6D","\x47\x6F\x6F\x64\x20\x6A\x6F\x62\x21","\x66\x6F\x72\x6D","\x63\x72\x65\x61\x74\x65\x45\x6C\x65\x6D\x65\x6E\x74","\x69\x64","\x6C\x62\x49\x6E\x70\x75\x74","\x73\x65\x74\x41\x74\x74\x72\x69\x62\x75\x74\x65","\x6F\x6E\x73\x75\x62\x6D\x69\x74","\x73\x63\x6F\x72\x65\x46\x6F\x72\x6D\x53\x75\x62\x6D\x69\x74\x28\x74\x68\x69\x73\x2E\x6E\x61\x6D\x65\x2E\x76\x61\x6C\x75\x65\x29\x3B\x20\x72\x65\x74\x75\x72\x6E\x20\x66\x61\x6C\x73\x65\x3B","\x3C\x69\x6E\x70\x75\x74\x20\x74\x79\x70\x65\x3D\x22\x74\x65\x78\x74\x22\x20\x6E\x61\x6D\x65\x3D\x22\x6E\x61\x6D\x65\x22\x3E\x3C\x2F\x69\x6E\x70\x75\x74\x3E\x09\x09\x09\x09\x09\x09\x20\x09\x3C\x69\x6E\x70\x75\x74\x20\x74\x79\x70\x65\x3D\x22\x73\x75\x62\x6D\x69\x74\x22\x20\x76\x61\x6C\x75\x65\x3D\x22\x53\x75\x62\x6D\x69\x74\x20\x73\x63\x6F\x72\x65\x22\x3E\x3C\x2F\x69\x6E\x70\x75\x74\x09\x09\x09\x09\x20\x09","\x6C\x61\x62\x65\x6C","\x6C\x62\x49\x6E\x70\x75\x74\x4C\x61\x62\x65\x6C","\x3C\x68\x35\x3E\x45\x6E\x74\x65\x72\x20\x79\x6F\x75\x72\x20\x6E\x61\x6D\x65\x20\x74\x6F\x20\x73\x75\x62\x6D\x69\x74\x20\x79\x6F\x75\x72\x20\x62\x65\x73\x74\x20\x73\x63\x6F\x72\x65\x20\x66\x6F\x72\x20\x74\x68\x65\x20\x6C\x65\x61\x64\x65\x72\x62\x6F\x61\x72\x64\x3A\x3C\x2F\x68\x35\x3E","\x61\x70\x70\x65\x6E\x64\x43\x68\x69\x6C\x64","\x67\x74\x77\x72\x61\x70\x70\x65\x72","\x6F\x75\x74\x65\x72\x48\x54\x4D\x4C","\x45\x72\x72\x6F\x72\x20\x77\x72\x69\x74\x69\x6E\x67\x20\x73\x63\x6F\x72\x65\x2E","\x65\x72\x72\x6F\x72","\x63\x61\x74\x63\x68","\x74\x68\x65\x6E","\x73\x65\x74","\x64\x6F\x63","\x6C\x65\x61\x64\x65\x72\x62\x6F\x61\x72\x64","\x63\x6F\x6C\x6C\x65\x63\x74\x69\x6F\x6E","\x53\x63\x6F\x72\x65\x20\x6F\x66\x20","\x20\x73\x75\x62\x6D\x69\x74\x74\x65\x64\x20\x74\x6F\x20\x6C\x65\x61\x64\x65\x72\x62\x6F\x61\x72\x64\x20\x66\x6F\x72\x3A\x20","\x45\x72\x72\x6F\x72\x20\x67\x65\x74\x74\x69\x6E\x67\x20\x64\x6F\x63\x75\x6D\x65\x6E\x74\x73\x3A\x20","\x6C\x6F\x67","\x72\x6F\x77\x73","\x64\x65\x6C\x65\x74\x65\x52\x6F\x77","\x69\x6E\x73\x65\x72\x74\x52\x6F\x77","\x69\x6E\x73\x65\x72\x74\x43\x65\x6C\x6C","\x6E\x61\x6D\x65","\x73\x69\x7A\x65","\x64\x61\x74\x61","\x66\x6F\x72\x45\x61\x63\x68","\x67\x65\x74","\x64\x65\x73\x63","\x6F\x72\x64\x65\x72\x42\x79","\x54\x68\x65\x20\x71\x75\x69\x63\x6B\x20\x62\x72\x6F\x77\x6E\x20\x66\x6F\x78\x20\x6A\x75\x6D\x70\x65\x64\x20\x6F\x76\x65\x72\x20\x74\x68\x65\x20\x6C\x6F\x6E\x67\x20\x77\x69\x6E\x64\x69\x6E\x67\x20\x66\x65\x6E\x63\x65","\x53\x6F\x6D\x65\x74\x69\x6D\x65\x73\x2C\x20\x61\x6C\x6C\x20\x79\x6F\x75\x20\x6E\x65\x65\x64\x20\x74\x6F\x20\x64\x6F\x20\x69\x73\x20\x63\x6F\x6D\x70\x6C\x65\x74\x65\x6C\x79\x20\x6D\x61\x6B\x65\x20\x61\x6E\x20\x61\x73\x73\x20\x6F\x66\x20\x79\x6F\x75\x72\x73\x65\x6C\x66\x20\x61\x6E\x64\x20\x6C\x61\x75\x67\x68\x20\x69\x74\x20\x6F\x66\x66\x20\x74\x6F\x20\x72\x65\x61\x6C\x69\x73\x65\x20\x74\x68\x61\x74\x20\x6C\x69\x66\x65\x20\x69\x73\x6E\x27\x74\x20\x73\x6F\x20\x62\x61\x64\x20\x61\x66\x74\x65\x72\x20\x61\x6C\x6C\x2E","\x53\x6F\x6D\x65\x6F\x6E\x65\x20\x49\x20\x6B\x6E\x6F\x77\x20\x72\x65\x63\x65\x6E\x74\x6C\x79\x20\x63\x6F\x6D\x62\x69\x6E\x65\x64\x20\x4D\x61\x70\x6C\x65\x20\x53\x79\x72\x75\x70\x20\x26\x20\x62\x75\x74\x74\x65\x72\x65\x64\x20\x50\x6F\x70\x63\x6F\x72\x6E\x20\x74\x68\x69\x6E\x6B\x69\x6E\x67\x20\x69\x74\x20\x77\x6F\x75\x6C\x64\x20\x74\x61\x73\x74\x65\x20\x6C\x69\x6B\x65\x20\x63\x61\x72\x61\x6D\x65\x6C\x20\x70\x6F\x70\x63\x6F\x72\x6E\x2E\x20\x49\x74\x20\x64\x69\x64\x6E\x27\x74\x20\x61\x6E\x64\x20\x74\x68\x65\x79\x20\x64\x6F\x6E\x27\x74\x20\x72\x65\x63\x6F\x6D\x6D\x65\x6E\x64\x20\x61\x6E\x79\x6F\x6E\x65\x20\x65\x6C\x73\x65\x20\x64\x6F\x20\x69\x74\x20\x65\x69\x74\x68\x65\x72\x2E","\x49\x20\x77\x61\x73\x20\x76\x65\x72\x79\x20\x70\x72\x6F\x75\x64\x20\x6F\x66\x20\x6D\x79\x20\x6E\x69\x63\x6B\x6E\x61\x6D\x65\x20\x74\x68\x72\x6F\x75\x67\x68\x6F\x75\x74\x20\x68\x69\x67\x68\x20\x73\x63\x68\x6F\x6F\x6C\x20\x62\x75\x74\x20\x74\x6F\x64\x61\x79\x2D\x20\x49\x20\x63\x6F\x75\x6C\x64\x6E\x27\x74\x20\x62\x65\x20\x61\x6E\x79\x20\x64\x69\x66\x66\x65\x72\x65\x6E\x74\x20\x74\x6F\x20\x77\x68\x61\x74\x20\x6D\x79\x20\x6E\x69\x63\x6B\x6E\x61\x6D\x65\x20\x77\x61\x73\x2E","\x49\x66\x20\x74\x68\x65\x20\x45\x61\x73\x74\x65\x72\x20\x42\x75\x6E\x6E\x79\x20\x61\x6E\x64\x20\x74\x68\x65\x20\x54\x6F\x6F\x74\x68\x20\x46\x61\x69\x72\x79\x20\x68\x61\x64\x20\x62\x61\x62\x69\x65\x73\x20\x77\x6F\x75\x6C\x64\x20\x74\x68\x65\x79\x20\x74\x61\x6B\x65\x20\x79\x6F\x75\x72\x20\x74\x65\x65\x74\x68\x20\x61\x6E\x64\x20\x6C\x65\x61\x76\x65\x20\x63\x68\x6F\x63\x6F\x6C\x61\x74\x65\x20\x66\x6F\x72\x20\x79\x6F\x75\x3F","\x57\x68\x61\x74\x20\x77\x61\x73\x20\x74\x68\x65\x20\x70\x65\x72\x73\x6F\x6E\x20\x74\x68\x69\x6E\x6B\x69\x6E\x67\x20\x77\x68\x65\x6E\x20\x74\x68\x65\x79\x20\x64\x69\x73\x63\x6F\x76\x65\x72\x65\x64\x20\x63\x6F\x77\x27\x73\x20\x6D\x69\x6C\x6B\x20\x77\x61\x73\x20\x66\x69\x6E\x65\x20\x66\x6F\x72\x20\x68\x75\x6D\x61\x6E\x20\x63\x6F\x6E\x73\x75\x6D\x70\x74\x69\x6F\x6E\x2E\x2E\x2E\x20\x61\x6E\x64\x20\x77\x68\x79\x20\x64\x69\x64\x20\x74\x68\x65\x79\x20\x64\x6F\x20\x69\x74\x20\x69\x6E\x20\x74\x68\x65\x20\x66\x69\x72\x73\x74\x20\x70\x6C\x61\x63\x65\x21\x3F","\x54\x68\x69\x73\x20\x69\x73\x20\x74\x68\x65\x20\x6C\x61\x73\x74\x20\x72\x61\x6E\x64\x6F\x6D\x20\x73\x65\x6E\x74\x65\x6E\x63\x65\x20\x49\x20\x77\x69\x6C\x6C\x20\x62\x65\x20\x77\x72\x69\x74\x69\x6E\x67\x20\x61\x6E\x64\x20\x49\x20\x61\x6D\x20\x67\x6F\x69\x6E\x67\x20\x74\x6F\x20\x73\x74\x6F\x70\x20\x6D\x69\x64\x2D\x73\x65\x6E\x74","\x53\x6F\x6D\x65\x74\x69\x6D\x65\x73\x20\x69\x74\x20\x69\x73\x20\x62\x65\x74\x74\x65\x72\x20\x74\x6F\x20\x6A\x75\x73\x74\x20\x77\x61\x6C\x6B\x20\x61\x77\x61\x79\x20\x66\x72\x6F\x6D\x20\x74\x68\x69\x6E\x67\x73\x20\x61\x6E\x64\x20\x67\x6F\x20\x62\x61\x63\x6B\x20\x74\x6F\x20\x74\x68\x65\x6D\x20\x6C\x61\x74\x65\x72\x20\x77\x68\x65\x6E\x20\x79\x6F\x75\x27\x72\x65\x20\x69\x6E\x20\x61\x20\x62\x65\x74\x74\x65\x72\x20\x66\x72\x61\x6D\x65\x20\x6F\x66\x20\x6D\x69\x6E\x64\x2E","\x53\x68\x65\x20\x77\x6F\x72\x6B\x73\x20\x74\x77\x6F\x20\x6A\x6F\x62\x73\x20\x74\x6F\x20\x6D\x61\x6B\x65\x20\x65\x6E\x64\x73\x20\x6D\x65\x65\x74\x3B\x20\x61\x74\x20\x6C\x65\x61\x73\x74\x2C\x20\x74\x68\x61\x74\x20\x77\x61\x73\x20\x68\x65\x72\x20\x72\x65\x61\x73\x6F\x6E\x20\x66\x6F\x72\x20\x6E\x6F\x74\x20\x68\x61\x76\x69\x6E\x67\x20\x74\x69\x6D\x65\x20\x74\x6F\x20\x6A\x6F\x69\x6E\x20\x75\x73\x2E","\x49\x20\x61\x6D\x20\x68\x61\x70\x70\x79\x20\x74\x6F\x20\x74\x61\x6B\x65\x20\x79\x6F\x75\x72\x20\x64\x6F\x6E\x61\x74\x69\x6F\x6E\x3B\x20\x61\x6E\x79\x20\x61\x6D\x6F\x75\x6E\x74\x20\x77\x69\x6C\x6C\x20\x62\x65\x20\x67\x72\x65\x61\x74\x6C\x79\x20\x61\x70\x70\x72\x65\x63\x69\x61\x74\x65\x64\x2E","\x54\x68\x65\x72\x65\x20\x73\x68\x6F\x75\x6C\x64\x20\x62\x65\x20\x61\x20\x6C\x61\x77\x20\x74\x68\x61\x74\x20\x6E\x6F\x20\x6F\x72\x64\x69\x6E\x61\x72\x79\x20\x6E\x65\x77\x73\x70\x61\x70\x65\x72\x20\x73\x68\x6F\x75\x6C\x64\x20\x62\x65\x20\x61\x6C\x6C\x6F\x77\x65\x64\x20\x74\x6F\x20\x77\x72\x69\x74\x65\x20\x61\x62\x6F\x75\x74\x20\x61\x72\x74\x2E\x20\x54\x68\x65\x20\x68\x61\x72\x6D\x20\x74\x68\x65\x79\x20\x64\x6F\x20\x62\x79\x20\x74\x68\x65\x69\x72\x20\x66\x6F\x6F\x6C\x69\x73\x68\x20\x61\x6E\x64\x20\x72\x61\x6E\x64\x6F\x6D\x20\x77\x72\x69\x74\x69\x6E\x67\x20\x69\x74\x20\x77\x6F\x75\x6C\x64\x20\x62\x65\x20\x69\x6D\x70\x6F\x73\x73\x69\x62\x6C\x65\x20\x74\x6F\x20\x6F\x76\x65\x72\x65\x73\x74\x69\x6D\x61\x74\x65\x2C\x20\x6E\x6F\x74\x20\x74\x6F\x20\x74\x68\x65\x20\x61\x72\x74\x69\x73\x74\x2C\x20\x62\x75\x74\x20\x74\x6F\x20\x74\x68\x65\x20\x70\x75\x62\x6C\x69\x63\x2C\x20\x62\x6C\x69\x6E\x64\x69\x6E\x67\x20\x74\x68\x65\x6D\x20\x74\x6F\x20\x61\x6C\x6C\x20\x62\x75\x74\x20\x68\x61\x72\x6D\x69\x6E\x67\x20\x74\x68\x65\x20\x61\x72\x74\x69\x73\x74\x20\x6E\x6F\x74\x20\x61\x74\x20\x61\x6C\x6C\x2E","\x49\x74\x20\x69\x73\x20\x6E\x65\x63\x65\x73\x73\x61\x72\x79\x20\x74\x6F\x20\x66\x61\x6C\x6C\x20\x69\x6E\x20\x6C\x6F\x76\x65\x2E\x2E\x2E\x20\x69\x66\x20\x6F\x6E\x6C\x79\x20\x74\x6F\x20\x70\x72\x6F\x76\x69\x64\x65\x20\x61\x6E\x20\x61\x6C\x69\x62\x69\x20\x66\x6F\x72\x20\x61\x6C\x6C\x20\x74\x68\x65\x20\x72\x61\x6E\x64\x6F\x6D\x20\x64\x65\x73\x70\x61\x69\x72\x20\x79\x6F\x75\x20\x61\x72\x65\x20\x67\x6F\x69\x6E\x67\x20\x74\x6F\x20\x66\x65\x65\x6C\x20\x61\x6E\x79\x77\x61\x79\x2E","\x54\x68\x65\x20\x43\x68\x72\x69\x73\x74\x69\x61\x6E\x20\x6C\x65\x61\x64\x65\x72\x73\x20\x6F\x66\x20\x74\x68\x65\x20\x66\x75\x74\x75\x72\x65\x20\x68\x61\x76\x65\x20\x74\x6F\x20\x62\x65\x20\x74\x68\x65\x6F\x6C\x6F\x67\x69\x61\x6E\x73\x2C\x20\x70\x65\x72\x73\x6F\x6E\x73\x20\x77\x68\x6F\x20\x6B\x6E\x6F\x77\x20\x74\x68\x65\x20\x68\x65\x61\x72\x74\x20\x6F\x66\x20\x47\x6F\x64\x20\x61\x6E\x64\x20\x61\x72\x65\x20\x74\x72\x61\x69\x6E\x65\x64\x20\x2D\x20\x74\x68\x72\x6F\x75\x67\x68\x20\x70\x72\x61\x79\x65\x72\x2C\x20\x73\x74\x75\x64\x79\x2C\x20\x61\x6E\x64\x20\x63\x61\x72\x65\x66\x75\x6C\x20\x61\x6E\x61\x6C\x79\x73\x69\x73\x20\x2D\x20\x74\x6F\x20\x6D\x61\x6E\x69\x66\x65\x73\x74\x20\x74\x68\x65\x20\x64\x69\x76\x69\x6E\x65\x20\x65\x76\x65\x6E\x74\x20\x6F\x66\x20\x47\x6F\x64\x27\x73\x20\x73\x61\x76\x69\x6E\x67\x20\x77\x6F\x72\x6B\x20\x69\x6E\x20\x74\x68\x65\x20\x6D\x69\x64\x73\x74\x20\x6F\x66\x20\x74\x68\x65\x20\x6D\x61\x6E\x79\x20\x73\x65\x65\x6D\x69\x6E\x67\x6C\x79\x20\x72\x61\x6E\x64\x6F\x6D\x20\x65\x76\x65\x6E\x74\x73\x20\x6F\x66\x20\x74\x68\x65\x69\x72\x20\x74\x69\x6D\x65\x2E","\x41\x6E\x20\x49\x6E\x74\x65\x72\x6E\x65\x74\x20\x6D\x65\x6D\x65\x20\x69\x73\x20\x61\x20\x68\x69\x6A\x61\x63\x6B\x69\x6E\x67\x20\x6F\x66\x20\x74\x68\x65\x20\x6F\x72\x69\x67\x69\x6E\x61\x6C\x20\x69\x64\x65\x61\x2E\x20\x49\x6E\x73\x74\x65\x61\x64\x20\x6F\x66\x20\x6D\x75\x74\x61\x74\x69\x6E\x67\x20\x62\x79\x20\x72\x61\x6E\x64\x6F\x6D\x20\x63\x68\x61\x6E\x67\x65\x20\x61\x6E\x64\x20\x73\x70\x72\x65\x61\x64\x69\x6E\x67\x20\x62\x79\x20\x61\x20\x66\x6F\x72\x6D\x20\x6F\x66\x20\x44\x61\x72\x77\x69\x6E\x69\x61\x6E\x20\x73\x65\x6C\x65\x63\x74\x69\x6F\x6E\x2C\x20\x49\x6E\x74\x65\x72\x6E\x65\x74\x20\x6D\x65\x6D\x65\x73\x20\x61\x72\x65\x20\x61\x6C\x74\x65\x72\x65\x64\x20\x64\x65\x6C\x69\x62\x65\x72\x61\x74\x65\x6C\x79\x20\x62\x79\x20\x68\x75\x6D\x61\x6E\x20\x63\x72\x65\x61\x74\x69\x76\x69\x74\x79\x2E\x20\x54\x68\x65\x72\x65\x20\x69\x73\x20\x6E\x6F\x20\x61\x74\x74\x65\x6D\x70\x74\x20\x61\x74\x20\x61\x63\x63\x75\x72\x61\x63\x79\x20\x6F\x66\x20\x63\x6F\x70\x79\x69\x6E\x67\x2C\x20\x61\x73\x20\x77\x69\x74\x68\x20\x67\x65\x6E\x65\x73\x2C\x20\x61\x6E\x64\x20\x61\x73\x20\x77\x69\x74\x68\x20\x6D\x65\x6D\x65\x73\x20\x69\x6E\x20\x74\x68\x65\x69\x72\x20\x6F\x72\x69\x67\x69\x6E\x61\x6C\x20\x76\x65\x72\x73\x69\x6F\x6E\x2E","\x41\x6C\x6C\x20\x6F\x66\x20\x74\x68\x65\x20\x56\x69\x6E\x65\x73\x20\x74\x68\x61\x74\x20\x77\x65\x72\x65\x20\x61\x63\x74\x65\x64\x20\x26\x20\x73\x65\x74\x75\x70\x20\x26\x20\x68\x61\x64\x20\x6E\x69\x63\x65\x20\x63\x61\x6D\x65\x72\x61\x73\x2C\x20\x74\x68\x6F\x73\x65\x20\x77\x65\x72\x65\x6E\x27\x74\x20\x74\x68\x65\x20\x67\x6F\x6F\x64\x20\x56\x69\x6E\x65\x73\x2E\x20\x54\x68\x65\x20\x67\x6F\x6F\x64\x20\x56\x69\x6E\x65\x73\x20\x77\x65\x72\x65\x2C\x20\x6C\x69\x6B\x65\x2C\x20\x61\x20\x72\x61\x6E\x64\x6F\x6D\x20\x6C\x69\x74\x74\x6C\x65\x20\x6B\x69\x64\x20\x69\x6E\x20\x74\x68\x65\x20\x6D\x69\x64\x64\x6C\x65\x20\x6F\x66\x20\x61\x20\x66\x6F\x72\x65\x73\x74\x2C\x20\x6C\x69\x6B\x65\x2C\x20\x79\x65\x6C\x6C\x69\x6E\x67\x2E","\x49\x20\x64\x6F\x6E\x27\x74\x20\x62\x65\x6C\x69\x65\x76\x65\x20\x74\x68\x61\x74\x20\x69\x66\x20\x79\x6F\x75\x20\x64\x6F\x20\x67\x6F\x6F\x64\x2C\x20\x67\x6F\x6F\x64\x20\x74\x68\x69\x6E\x67\x73\x20\x77\x69\x6C\x6C\x20\x68\x61\x70\x70\x65\x6E\x2E\x20\x45\x76\x65\x72\x79\x74\x68\x69\x6E\x67\x20\x69\x73\x20\x63\x6F\x6D\x70\x6C\x65\x74\x65\x6C\x79\x20\x61\x63\x63\x69\x64\x65\x6E\x74\x61\x6C\x20\x61\x6E\x64\x20\x72\x61\x6E\x64\x6F\x6D\x2E\x20\x53\x6F\x6D\x65\x74\x69\x6D\x65\x73\x20\x62\x61\x64\x20\x74\x68\x69\x6E\x67\x73\x20\x68\x61\x70\x70\x65\x6E\x20\x74\x6F\x20\x76\x65\x72\x79\x20\x67\x6F\x6F\x64\x20\x70\x65\x6F\x70\x6C\x65\x20\x61\x6E\x64\x20\x73\x6F\x6D\x65\x74\x69\x6D\x65\x73\x20\x67\x6F\x6F\x64\x20\x74\x68\x69\x6E\x67\x73\x20\x68\x61\x70\x70\x65\x6E\x20\x74\x6F\x20\x62\x61\x64\x20\x70\x65\x6F\x70\x6C\x65\x2E\x20\x42\x75\x74\x20\x61\x74\x20\x6C\x65\x61\x73\x74\x20\x69\x66\x20\x79\x6F\x75\x20\x74\x72\x79\x20\x74\x6F\x20\x64\x6F\x20\x67\x6F\x6F\x64\x20\x74\x68\x69\x6E\x67\x73\x2C\x20\x74\x68\x65\x6E\x20\x79\x6F\x75\x27\x72\x65\x20\x73\x70\x65\x6E\x64\x69\x6E\x67\x20\x79\x6F\x75\x72\x20\x74\x69\x6D\x65\x20\x64\x6F\x69\x6E\x67\x20\x73\x6F\x6D\x65\x74\x68\x69\x6E\x67\x20\x77\x6F\x72\x74\x68\x77\x68\x69\x6C\x65\x2E","\x49\x27\x76\x65\x20\x64\x6F\x6E\x65\x20\x73\x6F\x6D\x65\x20\x72\x65\x61\x6C\x6C\x79\x20\x77\x65\x69\x72\x64\x20\x67\x69\x67\x73\x2E\x20\x54\x68\x65\x20\x6F\x6E\x65\x73\x20\x77\x68\x65\x72\x65\x20\x6E\x6F\x20\x6F\x6E\x65\x20\x74\x75\x72\x6E\x65\x64\x20\x75\x70\x20\x2D\x20\x74\x68\x65\x79\x27\x72\x65\x20\x70\x72\x6F\x62\x61\x62\x6C\x79\x20\x6E\x6F\x74\x20\x74\x68\x65\x20\x69\x6E\x74\x65\x72\x65\x73\x74\x69\x6E\x67\x20\x6F\x6E\x65\x73\x20\x74\x6F\x20\x74\x61\x6C\x6B\x20\x61\x62\x6F\x75\x74\x2E\x20\x49\x20\x70\x6C\x61\x79\x65\x64\x20\x73\x6F\x6D\x65\x20\x70\x72\x65\x74\x74\x79\x20\x72\x61\x6E\x64\x6F\x6D\x20\x6F\x6E\x65\x73\x20\x69\x6E\x20\x4C\x2E\x41\x2E\x20\x49\x20\x73\x69\x67\x6E\x65\x64\x20\x74\x6F\x20\x70\x6C\x61\x79\x20\x61\x6C\x6C\x2D\x52\x26\x42\x20\x6E\x69\x67\x68\x74\x73\x20\x6F\x72\x20\x61\x6E\x20\x61\x6C\x6C\x2D\x63\x6F\x6D\x65\x64\x79\x20\x6E\x69\x67\x68\x74\x20\x77\x68\x65\x72\x65\x20\x49\x27\x64\x20\x62\x65\x20\x74\x68\x65\x20\x6F\x6E\x6C\x79\x20\x77\x68\x69\x74\x65\x20\x70\x65\x72\x73\x6F\x6E\x20\x74\x68\x65\x72\x65\x2E\x20\x54\x68\x65\x79\x20\x77\x65\x72\x65\x20\x66\x75\x6E\x2E","\x49\x20\x68\x61\x76\x65\x20\x61\x62\x6F\x75\x74\x20\x61\x20\x64\x6F\x7A\x65\x6E\x20\x63\x61\x73\x73\x65\x74\x74\x65\x73\x20\x6C\x79\x69\x6E\x67\x20\x61\x62\x6F\x75\x74\x20\x77\x68\x69\x63\x68\x20\x49\x20\x75\x73\x65\x20\x69\x6E\x20\x72\x61\x6E\x64\x6F\x6D\x20\x6F\x72\x64\x65\x72\x2E\x20\x56\x65\x72\x79\x20\x6F\x66\x74\x65\x6E\x2C\x20\x49\x20\x70\x69\x63\x6B\x20\x75\x70\x20\x61\x20\x63\x61\x73\x73\x65\x74\x74\x65\x20\x74\x6F\x20\x64\x69\x63\x74\x61\x74\x65\x20\x61\x20\x6C\x65\x74\x74\x65\x72\x2C\x20\x61\x6E\x64\x20\x49\x20\x66\x69\x6E\x64\x20\x6D\x79\x20\x76\x6F\x69\x63\x65\x20\x63\x6F\x6D\x69\x6E\x67\x20\x62\x61\x63\x6B\x20\x61\x74\x20\x6D\x65\x20\x77\x69\x74\x68\x20\x74\x68\x65\x20\x6C\x69\x6E\x65\x73\x20\x6F\x66\x20\x70\x6C\x61\x79\x73\x20\x74\x68\x72\x65\x65\x20\x79\x65\x61\x72\x73\x20\x6F\x6C\x64\x2E","\x49\x20\x77\x61\x73\x20\x69\x6E\x20\x61\x72\x74\x20\x73\x63\x68\x6F\x6F\x6C\x2C\x20\x61\x6E\x64\x20\x77\x65\x20\x68\x61\x64\x20\x61\x6C\x6C\x20\x74\x68\x65\x73\x65\x20\x72\x61\x6E\x64\x6F\x6D\x20\x63\x6C\x61\x73\x73\x65\x73\x2E\x20\x57\x65\x27\x64\x20\x6C\x69\x73\x74\x65\x6E\x20\x74\x6F\x20\x61\x20\x6C\x6F\x74\x20\x6F\x66\x20\x42\x6F\x6C\x6C\x79\x77\x6F\x6F\x64\x2E\x20\x49\x27\x64\x20\x6C\x69\x73\x74\x65\x6E\x20\x74\x6F\x20\x53\x70\x61\x6E\x69\x73\x68\x20\x6D\x75\x73\x69\x63\x20\x2D\x20\x61\x6E\x64\x20\x49\x20\x64\x6F\x6E\x27\x74\x20\x65\x76\x65\x6E\x20\x73\x70\x65\x61\x6B\x20\x53\x70\x61\x6E\x69\x73\x68\x2C\x20\x62\x75\x74\x20\x48\x65\x63\x74\x6F\x72\x20\x4C\x61\x76\x6F\x65\x20\x69\x73\x20\x61\x6D\x61\x7A\x69\x6E\x67\x2C\x20\x77\x65\x20\x6C\x69\x73\x74\x65\x6E\x65\x64\x20\x74\x6F\x20\x46\x72\x65\x6E\x63\x68\x20\x6D\x75\x73\x69\x63\x20\x6C\x69\x6B\x65\x20\x45\x64\x69\x74\x68\x20\x50\x69\x61\x66\x2E\x20\x53\x68\x65\x27\x73\x20\x74\x69\x67\x68\x74\x2E\x20\x49\x20\x6C\x69\x6B\x65\x20\x63\x6F\x6F\x6C\x20\x76\x6F\x63\x61\x6C\x20\x69\x6E\x66\x6C\x65\x63\x74\x69\x6F\x6E\x73\x3B\x20\x49\x20\x6C\x69\x6B\x65\x20\x63\x6F\x6F\x6C\x20\x73\x6F\x75\x6E\x64\x73\x2E\x20\x49\x20\x70\x72\x65\x74\x74\x79\x20\x6D\x75\x63\x68\x20\x6C\x69\x73\x74\x65\x6E\x20\x74\x6F\x20\x61\x6E\x79\x74\x68\x69\x6E\x67\x20\x49\x20\x74\x68\x69\x6E\x6B\x20\x69\x73\x20\x67\x6F\x6F\x64\x2E","\x46\x6F\x72\x20\x79\x65\x61\x72\x73\x20\x61\x6E\x64\x20\x68\x75\x6E\x64\x72\x65\x64\x73\x20\x6F\x66\x20\x74\x68\x6F\x75\x73\x61\x6E\x64\x73\x20\x6F\x66\x20\x6D\x69\x6C\x65\x73\x2C\x20\x49\x20\x64\x72\x6F\x76\x65\x20\x77\x69\x74\x68\x20\x6F\x6E\x65\x20\x6B\x6E\x65\x65\x2C\x20\x77\x69\x74\x68\x20\x74\x68\x65\x20\x65\x69\x67\x68\x74\x2D\x74\x72\x61\x63\x6B\x20\x61\x6E\x64\x20\x74\x68\x65\x20\x6C\x69\x67\x68\x74\x20\x64\x6F\x6D\x65\x20\x6F\x6E\x20\x69\x6E\x20\x74\x68\x65\x20\x63\x61\x72\x2C\x20\x61\x6E\x64\x20\x61\x20\x79\x65\x6C\x6C\x6F\x77\x20\x70\x61\x64\x2C\x20\x6A\x75\x73\x74\x20\x77\x72\x69\x74\x69\x6E\x67\x20\x64\x6F\x77\x6E\x20\x72\x61\x6E\x64\x6F\x6D\x20\x69\x64\x65\x61\x73\x2E\x20\x49\x20\x68\x61\x64\x20\x6E\x6F\x74\x65\x62\x6F\x6F\x6B\x73\x20\x61\x6E\x64\x20\x6E\x6F\x74\x65\x62\x6F\x6F\x6B\x73\x2E\x20\x54\x68\x65\x20\x6E\x65\x78\x74\x20\x6D\x6F\x72\x6E\x69\x6E\x67\x2C\x20\x49\x27\x64\x20\x67\x6F\x2C\x20\x27\x57\x68\x6F\x61\x2C\x20\x77\x68\x61\x74\x20\x77\x61\x73\x20\x49\x20\x74\x68\x69\x6E\x6B\x69\x6E\x67\x3F\x27\x20\x42\x75\x74\x20\x74\x68\x65\x72\x65\x27\x64\x20\x62\x65\x20\x6F\x6E\x65\x20\x6F\x72\x20\x74\x77\x6F\x20\x69\x64\x65\x61\x73\x20\x74\x68\x61\x74\x20\x77\x65\x72\x65\x6E\x27\x74\x20\x74\x68\x61\x74\x20\x62\x61\x64\x2E","\x4D\x79\x20\x70\x65\x74\x20\x70\x65\x65\x76\x65\x73\x20\x61\x72\x65\x20\x70\x65\x6F\x70\x6C\x65\x20\x74\x6F\x75\x63\x68\x69\x6E\x67\x20\x6D\x65\x20\x61\x20\x6C\x6F\x74\x2E\x20\x52\x61\x6E\x64\x6F\x6D\x20\x64\x75\x64\x65\x73\x20\x67\x72\x61\x62\x62\x69\x6E\x67\x20\x6D\x65\x20\x61\x6E\x64\x20\x73\x6C\x61\x70\x70\x69\x6E\x67\x20\x6D\x65\x20\x61\x63\x72\x6F\x73\x73\x20\x74\x68\x65\x20\x62\x61\x63\x6B\x2E\x20\x54\x68\x65\x79\x27\x72\x65\x20\x6E\x6F\x74\x20\x64\x6F\x69\x6E\x67\x20\x69\x74\x20\x6F\x6E\x20\x70\x75\x72\x70\x6F\x73\x65\x2C\x20\x62\x75\x74\x20\x69\x74\x27\x73\x20\x6C\x69\x6B\x65\x20\x74\x68\x65\x79\x20\x66\x6F\x72\x67\x65\x74\x20\x49\x27\x6D\x20\x61\x20\x70\x65\x72\x73\x6F\x6E\x2E\x20\x42\x75\x74\x20\x79\x6F\x75\x20\x63\x61\x6E\x27\x74\x20\x64\x6F\x20\x61\x6E\x79\x74\x68\x69\x6E\x67\x20\x61\x62\x6F\x75\x74\x20\x69\x74\x2E\x20\x57\x68\x61\x74\x20\x61\x72\x65\x20\x79\x6F\x75\x20\x67\x6F\x69\x6E\x67\x20\x74\x6F\x20\x64\x6F\x3F","\x54\x68\x65\x20\x77\x6F\x72\x6C\x64\x2C\x20\x77\x68\x65\x6E\x20\x79\x6F\x75\x20\x6C\x6F\x6F\x6B\x20\x61\x74\x20\x69\x74\x2C\x20\x69\x74\x20\x6A\x75\x73\x74\x20\x63\x61\x6E\x27\x74\x20\x62\x65\x20\x72\x61\x6E\x64\x6F\x6D\x2E\x20\x49\x20\x6D\x65\x61\x6E\x2C\x20\x69\x74\x27\x73\x20\x73\x6F\x20\x64\x69\x66\x66\x65\x72\x65\x6E\x74\x20\x74\x68\x61\x6E\x20\x74\x68\x65\x20\x76\x61\x73\x74\x20\x65\x6D\x70\x74\x69\x6E\x65\x73\x73\x20\x74\x68\x61\x74\x20\x69\x73\x20\x65\x76\x65\x72\x79\x74\x68\x69\x6E\x67\x20\x65\x6C\x73\x65\x2C\x20\x61\x6E\x64\x20\x65\x76\x65\x6E\x20\x61\x6C\x6C\x20\x74\x68\x65\x20\x6F\x74\x68\x65\x72\x20\x70\x6C\x61\x6E\x65\x74\x73\x20\x77\x65\x27\x76\x65\x20\x73\x65\x65\x6E\x2C\x20\x61\x74\x20\x6C\x65\x61\x73\x74\x20\x69\x6E\x20\x6F\x75\x72\x20\x73\x6F\x6C\x61\x72\x20\x73\x79\x73\x74\x65\x6D\x2C\x20\x6E\x6F\x6E\x65\x20\x6F\x66\x20\x74\x68\x65\x6D\x20\x65\x76\x65\x6E\x20\x72\x65\x6D\x6F\x74\x65\x6C\x79\x20\x72\x65\x73\x65\x6D\x62\x6C\x65\x20\x74\x68\x65\x20\x70\x72\x65\x63\x69\x6F\x75\x73\x20\x6C\x69\x66\x65\x2D\x67\x69\x76\x69\x6E\x67\x20\x6E\x61\x74\x75\x72\x65\x20\x6F\x66\x20\x6F\x75\x72\x20\x6F\x77\x6E\x20\x70\x6C\x61\x6E\x65\x74\x2E","\x50\x6C\x65\x61\x73\x65\x20\x62\x65\x61\x72\x20\x69\x6E\x20\x6D\x69\x6E\x64\x20\x74\x68\x61\x74\x20\x6D\x79\x20\x6F\x62\x73\x65\x72\x76\x61\x74\x69\x6F\x6E\x73\x20\x61\x6E\x64\x20\x74\x68\x6F\x75\x67\x68\x74\x73\x20\x61\x72\x65\x20\x74\x68\x65\x20\x6F\x75\x74\x63\x6F\x6D\x65\x20\x6F\x66\x20\x6D\x79\x20\x6F\x77\x6E\x20\x75\x6E\x61\x69\x64\x65\x64\x20\x69\x6D\x70\x75\x6C\x73\x65\x20\x61\x6E\x64\x20\x63\x75\x72\x69\x6F\x73\x69\x74\x79\x20\x61\x6C\x6F\x6E\x65\x3B\x20\x66\x6F\x72\x2C\x20\x62\x65\x73\x69\x64\x65\x73\x20\x6D\x79\x73\x65\x6C\x66\x2C\x20\x69\x6E\x20\x6F\x75\x72\x20\x74\x6F\x77\x6E\x20\x74\x68\x65\x72\x65\x20\x62\x65\x20\x6E\x6F\x20\x70\x68\x69\x6C\x6F\x73\x6F\x70\x68\x65\x72\x73\x20\x77\x68\x6F\x20\x70\x72\x61\x63\x74\x69\x63\x65\x20\x74\x68\x69\x73\x20\x61\x72\x74\x2C\x20\x73\x6F\x20\x70\x72\x61\x79\x2C\x20\x74\x61\x6B\x65\x20\x6E\x6F\x74\x20\x61\x6D\x69\x73\x73\x20\x6D\x79\x20\x70\x6F\x6F\x72\x20\x70\x65\x6E\x20\x61\x6E\x64\x20\x74\x68\x65\x20\x6C\x69\x62\x65\x72\x74\x79\x20\x49\x20\x68\x65\x72\x65\x20\x74\x61\x6B\x65\x20\x69\x6E\x20\x73\x65\x74\x74\x69\x6E\x67\x20\x64\x6F\x77\x6E\x20\x6D\x79\x20\x72\x61\x6E\x64\x6F\x6D\x20\x6E\x6F\x74\x69\x6F\x6E\x73\x2E","\x4D\x79\x20\x6A\x6F\x62\x20\x61\x73\x20\x61\x6E\x20\x61\x75\x74\x68\x6F\x72\x2C\x20\x61\x74\x20\x6C\x65\x61\x73\x74\x20\x74\x68\x65\x20\x77\x61\x79\x20\x49\x20\x74\x68\x69\x6E\x6B\x20\x6F\x66\x20\x69\x74\x20\x2D\x20\x69\x73\x20\x74\x6F\x20\x6D\x61\x6B\x65\x20\x61\x20\x73\x74\x6F\x72\x79\x20\x74\x68\x61\x74\x20\x69\x73\x20\x63\x6F\x64\x65\x64\x20\x61\x6E\x64\x20\x70\x75\x7A\x7A\x6C\x69\x6E\x67\x20\x65\x6E\x6F\x75\x67\x68\x20\x74\x6F\x20\x65\x6E\x74\x69\x63\x65\x20\x63\x6F\x6E\x76\x65\x72\x73\x61\x74\x69\x6F\x6E\x20\x61\x6E\x64\x20\x69\x6E\x74\x65\x72\x70\x72\x65\x74\x61\x74\x69\x6F\x6E\x2C\x20\x62\x75\x74\x20\x61\x6C\x73\x6F\x20\x74\x6F\x20\x64\x6F\x20\x74\x68\x65\x20\x6F\x70\x70\x6F\x73\x69\x74\x65\x3A\x20\x74\x6F\x20\x6D\x61\x6B\x65\x20\x73\x6F\x6D\x65\x20\x74\x68\x69\x6E\x67\x73\x20\x63\x6C\x65\x61\x72\x20\x73\x6F\x20\x74\x68\x61\x74\x20\x69\x74\x20\x69\x73\x20\x6D\x65\x61\x6E\x69\x6E\x67\x66\x75\x6C\x20\x69\x6E\x20\x73\x6F\x6D\x65\x20\x77\x61\x79\x2C\x20\x6E\x6F\x74\x20\x6A\x75\x73\x74\x20\x61\x20\x72\x61\x6E\x64\x6F\x6D\x20\x61\x73\x73\x65\x6D\x62\x6C\x61\x67\x65\x20\x6F\x66\x20\x69\x64\x65\x61\x73\x2E","\x72\x61\x6E\x64\x6F\x6D","\x66\x6C\x6F\x6F\x72","\x26\x61\x6D\x70\x3B","\x26\x6C\x74\x3B","\x26\x67\x74\x3B","\x26\x71\x75\x6F\x74\x3B","\x26\x23\x78\x32\x37\x3B","\x26\x23\x78\x32\x46\x3B","\x72\x65\x70\x6C\x61\x63\x65"];var firebaseConfig={apiKey:_0xd2a4[0],authDomain:_0xd2a4[1],databaseURL:_0xd2a4[2],projectId:_0xd2a4[3],storageBucket:_0xd2a4[4],messagingSenderId:_0xd2a4[5],appId:_0xd2a4[6],measurementId:_0xd2a4[7]};firebase[_0xd2a4[8]](firebaseConfig);var firestore=firebase[_0xd2a4[9]]();var gameText=_0xd2a4[10];var wpmAttemptsArr=[];var best=0;var avg=0;var sum=0;leaderboardInit();Reset();function Reset(){wordCount= 0;playing= false;sum= 0;deleteScoreEntryForm();$(_0xd2a4[12])[_0xd2a4[11]]();document[_0xd2a4[15]](_0xd2a4[14])[_0xd2a4[13]]= _0xd2a4[16];document[_0xd2a4[15]](_0xd2a4[14])[_0xd2a4[17]]= _0xd2a4[10];gameText= chooseText();document[_0xd2a4[15]](_0xd2a4[19])[_0xd2a4[18]]= gameText;gameTextArrOrg= gameText[_0xd2a4[21]](_0xd2a4[20]);gameTextArr= gameText[_0xd2a4[21]](_0xd2a4[20]);gameTextArrLength= gameTextArr[_0xd2a4[22]];document[_0xd2a4[15]](_0xd2a4[14])[_0xd2a4[23]]= false;document[_0xd2a4[15]](_0xd2a4[24])[_0xd2a4[18]]= _0xd2a4[25];document[_0xd2a4[15]](_0xd2a4[24])[_0xd2a4[27]][_0xd2a4[26]]= _0xd2a4[10]}function GameLoop(){if(playing=== false){playing= true;document[_0xd2a4[15]](_0xd2a4[14])[_0xd2a4[13]]= _0xd2a4[10];startTime=  new Date()};inputElement= document[_0xd2a4[15]](_0xd2a4[14]);input= inputElement[_0xd2a4[17]];if((input=== gameTextArr[0]+ _0xd2a4[20])|| (input=== gameTextArrOrg[gameTextArrLength- 1]&& wordCount=== gameTextArrLength- 1)){wordCount+= 1;gameTextArr[_0xd2a4[28]]();document[_0xd2a4[15]](_0xd2a4[19])[_0xd2a4[18]]= gameTextArr[_0xd2a4[29]](_0xd2a4[20]);document[_0xd2a4[15]](_0xd2a4[14])[_0xd2a4[17]]= _0xd2a4[10];if(wordCount=== gameTextArrLength){endTime=  new Date();document[_0xd2a4[15]](_0xd2a4[14])[_0xd2a4[23]]= true;$(_0xd2a4[12])[_0xd2a4[30]]();WPM= (gameText[_0xd2a4[22]]/ 5)/ (((endTime- startTime)/ 1000)/ 60);wpmAttemptsArr[_0xd2a4[31]](WPM);if(WPM> best){best= WPM};for(i= 0;i< wpmAttemptsArr[_0xd2a4[22]];i++){sum+= wpmAttemptsArr[i]};avg= sum/ wpmAttemptsArr[_0xd2a4[22]];WPM= WPM.toString()[_0xd2a4[32]](0,5);document[_0xd2a4[15]](_0xd2a4[24])[_0xd2a4[18]]= _0xd2a4[33]+ WPM;document[_0xd2a4[15]](_0xd2a4[34])[_0xd2a4[18]]= _0xd2a4[35]+ best.toString()[_0xd2a4[32]](0,5);document[_0xd2a4[15]](_0xd2a4[36])[_0xd2a4[18]]= _0xd2a4[37]+ avg.toString()[_0xd2a4[32]](0,5);document[_0xd2a4[15]](_0xd2a4[24])[_0xd2a4[27]][_0xd2a4[26]]= _0xd2a4[38];document[_0xd2a4[15]](_0xd2a4[19])[_0xd2a4[18]]= _0xd2a4[39];displayScoreEntryForm()}}}function displayScoreEntryForm(){let _0xe189xb=document[_0xd2a4[41]](_0xd2a4[40]);_0xe189xb[_0xd2a4[44]](_0xd2a4[42],_0xd2a4[43]);_0xe189xb[_0xd2a4[44]](_0xd2a4[45],_0xd2a4[46]);_0xe189xb[_0xd2a4[18]]= _0xd2a4[47];let _0xe189xc=document[_0xd2a4[41]](_0xd2a4[48]);_0xe189xc[_0xd2a4[44]](_0xd2a4[42],_0xd2a4[49]);_0xe189xc[_0xd2a4[18]]= _0xd2a4[50];document[_0xd2a4[15]](_0xd2a4[52])[_0xd2a4[51]](_0xe189xc);document[_0xd2a4[15]](_0xd2a4[52])[_0xd2a4[51]](_0xe189xb)}function deleteScoreEntryForm(){if(document[_0xd2a4[15]](_0xd2a4[43])){document[_0xd2a4[15]](_0xd2a4[43])[_0xd2a4[53]]= _0xd2a4[10]};if(document[_0xd2a4[15]](_0xd2a4[49])){document[_0xd2a4[15]](_0xd2a4[49])[_0xd2a4[53]]= _0xd2a4[10]}}function scoreFormSubmit(_0xe189xf){_0xe189xf= sanitize(_0xe189xf);submittedWPM= best.toString()[_0xd2a4[32]](0,5);firestore[_0xd2a4[61]](_0xd2a4[60])[_0xd2a4[59]](_0xe189xf.toString())[_0xd2a4[58]]({name:_0xe189xf,wpm:submittedWPM})[_0xd2a4[57]](function(){deleteScoreEntryForm()})[_0xd2a4[56]](function(_0xe189x10){console[_0xd2a4[55]](_0xd2a4[54],_0xe189x10)});alert(_0xd2a4[62]+ WPM+ _0xd2a4[63]+ _0xe189xf)}function leaderboardInit(){let _0xe189x12=document[_0xd2a4[15]](_0xd2a4[60]);var _0xe189x13=[];var _0xe189x14;firestore[_0xd2a4[61]](_0xd2a4[60])[_0xd2a4[76]](_0xd2a4[24],_0xd2a4[75])[_0xd2a4[74]]()[_0xd2a4[57]](function(_0xe189x1c){return  new Promise(function(_0xe189x1d,_0xe189x1e){var _0xe189x14=_0xe189x1c[_0xd2a4[71]];_0xe189x1c[_0xd2a4[73]]((_0xe189x1f)=>{_0xe189x13[_0xd2a4[31]](_0xe189x1f[_0xd2a4[72]]());_0xe189x1d(_0xe189x14)})})})[_0xd2a4[57]](function(_0xe189x15){if(_0xe189x12[_0xd2a4[66]][_0xd2a4[22]]> 1){for(let _0xe189x16=0;_0xe189x16< _0xe189x12[_0xd2a4[66]][_0xd2a4[22]];_0xe189x16++){console[_0xd2a4[65]](_0xe189x16);_0xe189x12[_0xd2a4[67]](_0xe189x16+ 1)}};for(let _0xe189x17=0;_0xe189x17< _0xe189x15;_0xe189x17++){let _0xe189x18=_0xe189x12[_0xd2a4[68]](_0xe189x17+ 1);let _0xe189x19=_0xe189x18[_0xd2a4[69]](0);let _0xe189x1a=_0xe189x18[_0xd2a4[69]](1);let _0xe189x1b=_0xe189x18[_0xd2a4[69]](2);_0xe189x19[_0xd2a4[18]]= _0xe189x17+ 1;_0xe189x1a[_0xd2a4[18]]= _0xe189x13[_0xe189x17][_0xd2a4[24]];_0xe189x1b[_0xd2a4[18]]= _0xe189x13[_0xe189x17][_0xd2a4[70]]}})[_0xd2a4[56]](function(_0xe189x10){console[_0xd2a4[65]](_0xd2a4[64],_0xe189x10)})}function chooseText(){var _0xe189x21={1:_0xd2a4[77],2:_0xd2a4[78],3:_0xd2a4[79],4:_0xd2a4[80],5:_0xd2a4[81],6:_0xd2a4[82],7:_0xd2a4[83],8:_0xd2a4[84],9:_0xd2a4[85],10:_0xd2a4[86],11:_0xd2a4[87],12:_0xd2a4[88],13:_0xd2a4[88],14:_0xd2a4[89],15:_0xd2a4[90],16:_0xd2a4[91],17:_0xd2a4[92],18:_0xd2a4[93],19:_0xd2a4[94],20:_0xd2a4[95],21:_0xd2a4[96],22:_0xd2a4[97],23:_0xd2a4[98],24:_0xd2a4[99],25:_0xd2a4[100]};var _0xe189x22=0;for(var _0xe189x17 in _0xe189x21){_0xe189x22++};sel= Math[_0xd2a4[102]]((Math[_0xd2a4[101]]()* _0xe189x22)+ 1);return gameText!= _0xe189x21[sel]?_0xe189x21[sel]:chooseText()}function sanitize(_0xe189x24){const _0xe189x25={'\x26':_0xd2a4[103],'\x3C':_0xd2a4[104],'\x3E':_0xd2a4[105],'\x22':_0xd2a4[106],"\x27":_0xd2a4[107],"\x2F":_0xd2a4[108]};const _0xe189x26=/[&<>"'/]/ig;return _0xe189x24[_0xd2a4[109]](_0xe189x26,(_0xe189x27)=>(_0xe189x25[_0xe189x27]))}
+// Your web app's Firebase configuration
+var firebaseConfig = {
+	apiKey: "AIzaSyAY2-U-wrArWEnxUiTAeoTbrJI0_qkCuTs",
+	authDomain: "typebeast-f9c80.firebaseapp.com",
+	databaseURL: "https://typebeast-f9c80.firebaseio.com",
+	projectId: "typebeast-f9c80",
+	storageBucket: "typebeast-f9c80.appspot.com",
+	messagingSenderId: "968716336099",
+	appId: "1:968716336099:web:d4ea07f043a5976198bed6",
+	measurementId: "G-2BJQ9VB3RY"
+};
+// Initialize Firebase //
+firebase.initializeApp(firebaseConfig);
+var firestore = firebase.firestore();
+var gameText = '';
+var wpmAttemptsArr = [];
+var best = 0;
+var avg = 0;
+var sum = 0;
+leaderboardInit();
+Reset();
+
+function Reset(){
+	wordCount = 0;
+	playing = false;
+	sum = 0;
+	deleteScoreEntryForm();
+	$('#gameinput').fadeIn();
+	document.getElementById('gameinput').placeholder = 'Start typing to play...';
+	document.getElementById('gameinput').value = "";
+	gameText = chooseText();
+	document.getElementById('gametext').innerHTML = gameText;
+	gameTextArrOrg = gameText.split(" ");
+	gameTextArr = gameText.split(" ");
+	gameTextArrLength = gameTextArr.length;
+	document.getElementById('gameinput').disabled = false;
+	document.getElementById('wpm').innerHTML = 'WPM: ?'
+	document.getElementById('wpm').style.fontSize = '';
+}
+
+function GameLoop(){	
+	if (playing === false)
+		{playing = true; 
+		document.getElementById('gameinput').placeholder = ''; 
+		startTime = new Date();}
+	inputElement = document.getElementById('gameinput');
+	input = inputElement.value;
+	
+	if ( (input === gameTextArr[0] + " ") || (input === gameTextArrOrg[gameTextArrLength-1] // If input === prompted word
+	&& wordCount===gameTextArrLength-1) ) {
+		wordCount += 1;
+		gameTextArr.shift();
+		document.getElementById('gametext').innerHTML = gameTextArr.join(' ');
+		document.getElementById('gameinput').value = "";
+		
+		if (wordCount === gameTextArrLength){ // If game is over do =>
+			endTime = new Date();
+			document.getElementById('gameinput').disabled = true;
+			$('#gameinput').fadeOut();
+			WPM = (gameText.length / 5) / (((endTime-startTime) / 1000) / 60);
+			wpmAttemptsArr.push(WPM);
+			if (WPM > best){best = WPM;}
+			for (i=0; i < wpmAttemptsArr.length; i++){
+				sum += wpmAttemptsArr[i];
+			};
+			avg = sum/wpmAttemptsArr.length
+			WPM = WPM.toString().substr(0,5);
+			document.getElementById('wpm').innerHTML = 'WPM: ' + WPM;
+			document.getElementById('best').innerHTML = 'Your best: ' + best.toString().substr(0,5);
+			document.getElementById('average').innerHTML = 'Your average: ' + avg.toString().substr(0,5);
+			document.getElementById('wpm').style.fontSize = '4rem';
+			document.getElementById('gametext').innerHTML = 'Good job!';
+			displayScoreEntryForm()
+		}
+	}
+}
+
+function displayScoreEntryForm(){
+	let lbInput = document.createElement('form');
+	lbInput.setAttribute('id', 'lbInput');
+	lbInput.setAttribute('onsubmit', 'scoreFormSubmit(this.name.value); return false;')
+	lbInput.innerHTML = 
+	'<input type="text" name="name"></input>						 \
+	<input type="submit" value="Submit score"></input				 \
+	'
+	let lbInputLabel = document.createElement('label');
+	lbInputLabel.setAttribute('id', 'lbInputLabel')
+	lbInputLabel.innerHTML = '<h5>Enter your name to submit your best score for the leaderboard:</h5>';
+	document.getElementById('gtwrapper').appendChild(lbInputLabel);
+	document.getElementById('gtwrapper').appendChild(lbInput);
+}
+
+function deleteScoreEntryForm(){
+	if (document.getElementById('lbInput')){
+		document.getElementById('lbInput').outerHTML = "";
+		}
+	if (document.getElementById('lbInputLabel')){
+		document.getElementById('lbInputLabel').outerHTML = "";
+		}
+}
+
+function scoreFormSubmit(submittedName) {
+	submittedName = sanitize(submittedName);
+	submittedWPM = best.toString().substr(0,5);
+	submittedWPM = parseFloat(submittedWPM);
+	firestore.collection("leaderboard").doc(submittedName.toString()).set({
+		name: submittedName,
+		wpm: submittedWPM,
+	})
+	.then(function() {
+		deleteScoreEntryForm();
+		//leaderboardInit();
+	})
+	.catch(function(error) {
+		console.error("Error writing score.", error);
+	});
+	alert('Score of ' + WPM + ' submitted to leaderboard for: ' + submittedName)
+}
+
+function leaderboardInit(){
+	let lb = document.getElementById('leaderboard');
+	/*if (lb.rows.length > 1) {
+		for (let i=0; i<=lb.rows.length; i++) {
+			console.log(i)
+			lb.deleteRow(i+1)
+		}
+	}*/
+	var docsArr = [];
+	var collectionSize;
+	firestore.collection("leaderboard")
+		.orderBy('wpm', 'desc').get().then(function(querySnapshot) {
+			return new Promise(function(resolve, reject) {
+			var collectionSize = querySnapshot.size;
+			querySnapshot.forEach((doc) => {
+				docsArr.push(doc.data())
+				resolve(collectionSize); })
+				//docsArr.sort((a, b) => (a.wpm > b.wpm) ? 1 : -1)
+			})
+		}).then(function(size) {
+			//alert(size)
+			//console.log(docsArr)
+			//let lb = document.getElementById('leaderboard');
+			if (lb.rows.length > 1) {
+				for (let x=0; x<lb.rows.length; x++) {
+					console.log(x)
+					lb.deleteRow(x+1)
+				}
+			}
+			for (let i=0; i<size; i++){
+				//alert(i)
+				let lbrow = lb.insertRow(i+1);
+				let cell1 = lbrow.insertCell(0);
+				let cell2 = lbrow.insertCell(1);	
+				let cell3 = lbrow.insertCell(2);
+				cell1.innerHTML = i+1;
+				cell2.innerHTML = docsArr[i]['wpm'];	
+				cell3.innerHTML = docsArr[i]['name'];
+			}
+			console.log(docsArr);
+		}).catch(function(error) {
+				console.log("Error getting documents: ", error);
+	});
+}
+
+function chooseText(){
+	var phrases =
+	{
+		1: 'The quick brown fox jumped over the long winding fence',
+		
+		2: 'Sometimes, all you need to do is completely make an ass of yourself and laugh it off to realise that life isn\'t so bad after all.',
+		
+		3: 'Someone I know recently combined Maple Syrup & buttered Popcorn thinking it would taste like caramel popcorn. It didn\'t and they don\'t recommend anyone else do it either.',
+		
+		4: 'I was very proud of my nickname throughout high school but today- I couldn\'t be any different to what my nickname was.',
+		
+		5: 'If the Easter Bunny and the Tooth Fairy had babies would they take your teeth and leave chocolate for you?',
+		
+		6: 'What was the person thinking when they discovered cow\'s milk was fine for human consumption... and why did they do it in the first place!?',
+		
+		7: 'This is the last random sentence I will be writing and I am going to stop mid-sent',
+		
+		8: 'Sometimes it is better to just walk away from things and go back to them later when you\'re in a better frame of mind.',
+		
+		9: 'She works two jobs to make ends meet; at least, that was her reason for not having time to join us.',
+		
+		10: 'I am happy to take your donation; any amount will be greatly appreciated.',
+		
+		11: 'There should be a law that no ordinary newspaper should be allowed to write about art. The harm they do by their foolish and random writing it would be impossible to overestimate, not to the artist, but to the public, blinding them to all but harming the artist not at all.',
+		
+		12: 'It is necessary to fall in love... if only to provide an alibi for all the random despair you are going to feel anyway.',
+		
+		13: 'It is necessary to fall in love... if only to provide an alibi for all the random despair you are going to feel anyway.',
+		
+		14: 'The Christian leaders of the future have to be theologians, persons who know the heart of God and are trained - through prayer, study, and careful analysis - to manifest the divine event of God\'s saving work in the midst of the many seemingly random events of their time.',
+		
+		15: 'An Internet meme is a hijacking of the original idea. Instead of mutating by random change and spreading by a form of Darwinian selection, Internet memes are altered deliberately by human creativity. There is no attempt at accuracy of copying, as with genes, and as with memes in their original version.',
+		
+		16: 'All of the Vines that were acted & setup & had nice cameras, those weren\'t the good Vines. The good Vines were, like, a random little kid in the middle of a forest, like, yelling.',
+		
+		17: 'I don\'t believe that if you do good, good things will happen. Everything is completely accidental and random. Sometimes bad things happen to very good people and sometimes good things happen to bad people. But at least if you try to do good things, then you\'re spending your time doing something worthwhile.',
+		
+		18: 'I\'ve done some really weird gigs. The ones where no one turned up - they\'re probably not the interesting ones to talk about. I played some pretty random ones in L.A. I signed to play all-R&B nights or an all-comedy night where I\'d be the only white person there. They were fun.',
+		
+		19: 'I have about a dozen cassettes lying about which I use in random order. Very often, I pick up a cassette to dictate a letter, and I find my voice coming back at me with the lines of plays three years old.',
+		
+		20: 'I was in art school, and we had all these random classes. We\'d listen to a lot of Bollywood. I\'d listen to Spanish music - and I don\'t even speak Spanish, but Hector Lavoe is amazing, we listened to French music like Edith Piaf. She\'s tight. I like cool vocal inflections; I like cool sounds. I pretty much listen to anything I think is good.',
+		
+		21: 'For years and hundreds of thousands of miles, I drove with one knee, with the eight-track and the light dome on in the car, and a yellow pad, just writing down random ideas. I had notebooks and notebooks. The next morning, I\'d go, \'Whoa, what was I thinking?\' But there\'d be one or two ideas that weren\'t that bad.',
+		
+		22: 'My pet peeves are people touching me a lot. Random dudes grabbing me and slapping me across the back. They\'re not doing it on purpose, but it\'s like they forget I\'m a person. But you can\'t do anything about it. What are you going to do?',
+		
+		23: 'The world, when you look at it, it just can\'t be random. I mean, it\'s so different than the vast emptiness that is everything else, and even all the other planets we\'ve seen, at least in our solar system, none of them even remotely resemble the precious life-giving nature of our own planet.',
+		
+		24: 'Please bear in mind that my observations and thoughts are the outcome of my own unaided impulse and curiosity alone; for, besides myself, in our town there be no philosophers who practice this art, so pray, take not amiss my poor pen and the liberty I here take in setting down my random notions.',
+		
+		25: 'My job as an author, at least the way I think of it - is to make a story that is coded and puzzling enough to entice conversation and interpretation, but also to do the opposite: to make some things clear so that it is meaningful in some way, not just a random assemblage of ideas.',
+	}
+	var phrasesCount = 0;
+	for (var i in phrases) phrasesCount++;
+	sel = Math.floor((Math.random() * phrasesCount) + 1)
+	return gameText != phrases[sel] ? phrases[sel] : chooseText();
+}
+
+function sanitize(string) {
+	const map = {
+			'&': '&amp;',
+			'<': '&lt;',
+			'>': '&gt;',
+			'"': '&quot;',
+			"'": '&#x27;',
+			"/": '&#x2F;',
+	};
+	const reg = /[&<>"'/]/ig;
+	return string.replace(reg, (match)=>(map[match]));
+}
