--- main.js.tmp.html
@@ -1,1 +1,0 @@
-<script src="file:///Users/jacobcritch/Documents/TypingTest/main.js"></script>
