--- main_source.js.tmp.html
@@ -1,1 +1,0 @@
-<script src="file:///Users/jacobcritch/Documents/TypingTest/main_source.js"></script>
